Changed double declarations to shorthand, changed overridden method javadocs to inherit parent docs

+++ src/filters/MaxFilterNTest.java
@@ -20,31 +20,31 @@ public class MaxFilterNTest {
 	
 	@Test
 	public void testInputOutput() throws InvalidObjectException {
-		Double input = (double) 2;
+		Double input = 2d;
 		MaxFilter filter = new MaxFilterN(2);
 		
 		filter.addInput(input);
 		assertEquals("Only value should be output", input, filter.getOutput());
 		
-		input = (double) 3;
+		input = 3d;
 		filter.addInput(input);
 		assertEquals("New value should be output", input, filter.getOutput());
 		
-		input = (double) 1;
+		input = 1d;
 		filter.addInput(input);
 		assertNotEquals("Old value should still be output", input, filter.getOutput());
 		
-		input = (double) 2;
+		input = 2d;
 		filter.addInput(input);
 		assertEquals("Old max should be popped due to age", input, filter.getOutput());
 	}
 
 	@Test
 	public void testReset() throws InvalidObjectException {
-		Double reset = (double) 0;
+		Double reset = 0d;
 		MaxFilterN filter = new MaxFilterN(2);
 		
-		filter.addInput((double) 27);
+		filter.addInput(27d);
 		filter.reset(reset);
 		
 		assertEquals("Output should be the reset value", reset, filter.getOutput());
